requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
Flask==2.2.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
logging==0.4.9.6
MarkupSafe==3.0.2
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pyTelegramBotAPI==4.8.0
requests==2.32.3
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3